backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
aiomysql==0.2.0
pymysql==1.1.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
Pillow==10.4.0
aiofiles==24.1.0
websockets==12.0
python-dotenv==1.0.1
alembic==1.13.0
cryptography==43.0.0
